pyproject.toml
[project]
name = "simulation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "executing==2.2.0",
    "filetype==1.2.0",
    "fonttools==4.58.2",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.0",
    "google-auth==2.40.3",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.3",
    "grpcio==1.72.1",
    "grpcio-status==1.72.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kiwisolver==1.4.8",
    "langchain==0.3.25",
    "langchain-core==0.3.64",
    "langchain-google-genai==2.1.5",
    "langchain-text-splitters==0.3.8",
    "langsmith==0.3.45",
    "matplotlib==3.10.3",
    "matplotlib-inline==0.1.7",
    "nest-asyncio==1.6.0",
    "numpy==2.2.6",
    "orjson==3.10.18",
    "packaging==24.2",
    "pandas==2.3.0",
    "parso==0.8.4",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "pywin32==310",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "stack-data==0.6.3",
    "tabulate==0.9.0",
    "tenacity==9.1.2",
    "tornado==6.5.1",
    "traitlets==5.14.3",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
    "zstandard==0.23.0",
]
